refactor: use ObservableInputTuple in xWith ops
