+++ spec/operators/zipWith-spec.ts
@@ -192,25 +192,6 @@ describe('zipWith', () => {
         expectSubscriptions(a.subscriptions).toBe(asubs);
       });
     });
-
-    it('should work with non-empty observable and non-empty iterable selector that throws', () => {
-      rxTestScheduler.run(({ hot, expectObservable, expectSubscriptions }) => {
-        const a = hot('---^--1--2--3--|');
-        const asubs = '   ^-----!';
-        const expected = '---x--#';
-        const b = [4, 5, 6];
-
-        const selector = function(x: string, y: number) {
-          if (y === 5) {
-            throw new Error('too bad');
-          } else {
-            return x + y;
-          }
-        };
-        expectObservable(a.pipe(zipWith(b, selector))).toBe(expected, { x: '14' }, new Error('too bad'));
-        expectSubscriptions(a.subscriptions).toBe(asubs);
-      });
-    });
   });
 
   it('should work with n-ary symmetric', () => {
